requirements.txt
langchain
camelot-py
dashscope==1.22.1
ipdb==0.13.13
numpy==1.26.4
pdfplumber==0.9.0
pillow==11.0.0
opencv-python==4.11.0.86
langchain_community==0.3.17
langchain_core==0.3.34
langchain_text_splitters==0.3.6
langgraph==0.2.70
pandas==2.2.3
pytesseract==0.3.13
requests==2.32.3
